refactored sigmoid using the new interface
